roles/test-profiles: clean previous failed install log
By default, phoronix-test-suites do not remove logs of previous failed
test-profile installation even if later it installed it successfully.
This could cause a false positive as the error log file is used to
trigger if the installation of the test-profile is successful or not.

To fix this, always check if this log file exists, and remove it.

Signed-off-by: Paul Le Guen de Kerneizon <paul.leguendekerneizon@savoirfairelinux.com>