run_test_profiles: rename test_profile variable
This variable is passed to the test profiles roles and tell which one
have to be run (monitoring, benchmark).
As this variable can be defined as a list in the inventory (if it
includes test profiles arguments for example), it is better to rename
it to test_profiles to avoid confusion.

Signed-off-by: Paul Le Guen de Kerneizon <paul.leguendekerneizon@savoirfairelinux.com>