roles: set unique test-results name
Currently, the test-results directory name generated by the
phoronix-test-suite is a the timestamp of the start time of the test.
This behavior makes difficult to identify the test-results directory
when multiple tests are run in parallel on a target.

To fix this, this commit sets a unique identifer for the test results
directory, composed of the test name and the timestamp of the start
time.

Signed-off-by: Paul Le Guen de Kerneizon <paul.leguendekerneizon@savoirfairelinux.com>