run_monitoring_test_profiles: pass test profiles arguments through vars
Currently, test profiles arguments are passe to the
phoronix-test-suite using a template of the test-definition.xml
file, with a set of this value from the inventory file.

This is not ideal as:
- It needs to perform a new configuration when one argument of the
test profile is changed
- It raises a warning when the test-definition.xml file is not
templated (e.g. is the test profile is not using arguments), which
can confuse the user.

To fix this, this commit passes the test profiles arguments through shell
environment variables. These arguments must be set in the
vars/main.yaml file of the run benchmark test profiles role.

Signed-off-by: Paul Le Guen de Kerneizon <paul.leguendekerneizon@savoirfairelinux.com>