get_monitoring_test_results: improve test_run_id identification
Currently, the test-profile run is identified by a variable
`test_run_id` composed of the test-profile name and the start time of
the test. The current implementation is not robust enough, and
overwritten can occurs if two test-profiles are run at the same time.

To fix this, this variable is now built in the
`run_monitoring_tests_profiles` role, and is then passed to the
`get_monitoring_test_results` role for each monitoring test profile
run.

Signed-off-by: Paul Le Guen de Kerneizon <paul.leguendekerneizon@savoirfairelinux.com>