get_monitoring_test_results/process_monitoring: fix incorrect wait
When ending process_monitoring test-profile we need to ensure it has
done all remaning work (results generation), before doing any
post-processing.
Currently, the wait task used is incorrect, and post-processing is
done before the end of the test-profile. This result in missing pie
plot generation in process_monitoring report.

To fix this, this commit records the active monitoring test-profile in
execution in an Ansible async_status variable, a wait for it to finish.

Signed-off-by: Paul Le Guen de Kerneizon <paul.leguendekerneizon@savoirfairelinux.com>