test-profiles/monitoring: gather monitoring test profile vars
Currently, multiples variables are used for each monitoring
test-profiles (test-profiles identification, test-profile name, etc.).
With the growing of the project, it is getting more and more difficult
and readable to understand what and which variable does.
To solve this, this commit regroups in one variable `active_monitoring`
the list of current active monitoring with:
- The test-profile name
- The Ansible status of the test-profile if it has been launched in
background
- The arguments used for this test-profile
- The identification of the test-profile run

Signed-off-by: Paul Le Guen de Kerneizon <paul.leguendekerneizon@savoirfairelinux.com>