run_benchmark_test_profiles: add support of menu entry
A test-profile can be run along many different options. For example,
`disk` test profile can be executed in time based mode (test is run for
a certain amount of time) or size based mode (test is run until a specific
amount of data has been written).
Phoronix makes the selection of this mode through a menu entry, which
user select which option he want to run for the test profile.

This commit adds the support of the automation of the selected
test-profiles.

Signed-off-by: Paul Le Guen de Kerneizon <paul.leguendekerneizon@savoirfairelinux.com>