run_benchmark_test_profiles: reset test_profile_args var
A var with the same name is used in run_monitoring_tests_profiles role.
Ansible do not reset the value of this var accross roles, and this is
causing a contamination of monitoring test profile arguments on
benchmark test profiles.
To fix this, force the value of this var to none.

Signed-off-by: Paul Le Guen de Kerneizon <paul.leguendekerneizon@savoirfairelinux.com>