run_monitoring_test_profiles: reset test_profile_args var
A var with the same name is used in run_monitoring_tests_profiles role.
Ansible do not reset the value of this var accross roles, and this is
causing a contamination of one monitoring test profile arguments to
another monitoring test profile.
To fix this, force the value of this var to none.

Signed-off-by: Paul Le Guen de Kerneizon <paul.leguendekerneizon@savoirfairelinux.com>