run_benchmark_test_profiles: fix incorrect test arguments retrieve
When test arguments are not specified from test scenario
file, they must be retrieved from run_benchmark_test_profiles role var
file.
In this case the value of the `test_profile_args` is `none`. This
commit fixes incorrect and missing `when` conditions. Plus, it retrieves
only the first argument defined in the test profile vars.

Signed-off-by: Paul Le Guen de Kerneizon <paul.leguendekerneizon@savoirfairelinux.com>