Dockerfile: remove ubuntu user
Ubuntu 24.04 Docker image creates a `ubuntu` user with uuid 1000. This
user is then used by cqfd, instead of the current user who launched the
cqfd command.
This can cause issues with SSH permission, specially if SSH keys are
not installed in ~/.ssh directory.
To fix this, this commit removes this user, in order to to use the
current user who launched the cqfd command.

Signed-off-by: Paul Le Guen de Kerneizon <paul.leguendekerneizon@savoirfairelinux.com>